Package web files into package.
To achieve this, copy the files over and swap to using a runtime
find of the web files, via an environment variable. This env var is set
during the startup of the server, which should know the install location
of the package (whereas knowing the install location pre-install seems
iffy).